Move collision exclusion functions into separate files.
