+++ CMakeLists.txt
@@ -93,6 +93,7 @@ add_library(Edyn
     src/edyn/shapes/paged_triangle_mesh.cpp
     src/edyn/util/triangle_util.cpp
     src/edyn/util/ragdoll.cpp
+    src/edyn/util/exclude_collision.cpp
     src/edyn/shapes/box_shape.cpp
     src/edyn/shapes/cylinder_shape.cpp
     src/edyn/shapes/polyhedron_shape.cpp
